+++ packages/loader/CHANGELOG.md
@@ -5,3 +5,9 @@
 ### Major Changes
 
 - ESM modules
+
+## 0.6.0
+
+### Minor Changes
+
+- upgrade dependencies & syntax