+++ packages/validator-cli/CHANGELOG.md
@@ -6,8 +6,16 @@
 
 - ESM modules
 
+## 0.6.0
+
+### Minor Changes
+
+- upgrade dependencies & syntax
+
 ### Patch Changes
 
 - Updated dependencies
   - typeonly@1.0.0
   - @typeonly/validator@1.0.0
+  - @typeonly/validator@0.6.0
+  - typeonly@0.5.0