+++ packages/validator/CHANGELOG.md
@@ -6,6 +6,12 @@
 
 - ESM modules
 
+## 0.6.0
+
+### Minor Changes
+
+- upgrade dependencies & syntax
+
 ### Patch Changes
 
 - Updated dependencies
@@ -16,3 +22,4 @@
 ### Patch Changes
 
 - Improve validation error message (fix)
+  - @typeonly/loader@0.6.0